# src/backend/openapi.yaml
openapi: 3.1.0
info:
  title: LiDAR Forest Analysis API
  version: 1.0.0
  description: |
    REST API for the LiDAR Forest Analysis Platform. Transform LiDAR point cloud data
    into professional forest inventory reports, carbon stock estimates, and 3D visualizations.

    ## Authentication

    All API requests require authentication via API key. Include your API key in requests using
    one of these methods:

    - **Authorization header** (recommended): `Authorization: Bearer lf_live_xxx`
    - **X-API-Key header**: `X-API-Key: lf_live_xxx`

    ## Rate Limits

    Rate limits vary by tier:

    | Tier | Requests/Minute | Requests/Day |
    |------|-----------------|--------------|
    | Free | 60 | 10,000 |
    | Starter | 120 | 50,000 |
    | Professional | 300 | 200,000 |
    | Enterprise | 1,000 | Unlimited |

    Rate limit headers are included in all responses:
    - `X-RateLimit-Limit`: Max requests per minute
    - `X-RateLimit-Remaining`: Remaining requests this minute
    - `X-RateLimit-Reset`: Time when the rate limit resets (ISO 8601)
  contact:
    name: API Support
    email: api@lidarforest.com
    url: https://docs.lidarforest.com
  license:
    name: Proprietary
    url: https://lidarforest.com/terms
servers:
  - url: https://api.lidarforest.com/api/v1
    description: Production server
  - url: https://staging-api.lidarforest.com/api/v1
    description: Staging server
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /reports:
    get:
      summary: List reports
      tags: [Reports]
      operationId: listReports
      parameters:
        - name: projectId
          in: query
          schema:
            type: string
            format: uuid
        - name: analysisId
          in: query
          schema:
            type: string
            format: uuid
        - name: type
          in: query
          schema:
            $ref: '#/components/schemas/ReportType'
        - name: format
          in: query
          schema:
            $ref: '#/components/schemas/ReportFormat'
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of reports
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportListResponse'
    post:
      summary: Generate report
      tags: [Reports]
      operationId: createReport
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReportInput'
      responses:
        '201':
          description: Report generation started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportResponse'
  /reports/{reportId}:
    get:
      summary: Get report details
      tags: [Reports]
      operationId: getReport
      parameters:
        - $ref: '#/components/parameters/ReportId'
      responses:
        '200':
          description: Report details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportDetailResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Delete report
      tags: [Reports]
      operationId: deleteReport
      parameters:
        - $ref: '#/components/parameters/ReportId'
      responses:
        '200':
          description: Report deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /reports/{reportId}/download:
    get:
      summary: Download report
      tags: [Reports]
      operationId: downloadReport
      parameters:
        - $ref: '#/components/parameters/ReportId'
      responses:
        '200':
          description: Download URL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportDownloadResponse'
        '410':
          description: Report expired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: API key as Bearer token
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key header
  parameters:
...
    ReportId:
      name: reportId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
...
    ReportType:
      type: string
      enum: [INVENTORY, CARBON, TIMBER_VALUE, GROWTH_PROJECTION, FULL]
    ReportFormat:
      type: string
      enum: [PDF, EXCEL, CSV, JSON]
...
    CreateReportInput:
      type: object
      required: [analysisId, name, type]
      properties:
        analysisId:
          type: string
          format: uuid
        name:
          type: string
          minLength: 1
          maxLength: 200
        type:
          $ref: '#/components/schemas/ReportType'
        format:
          $ref: '#/components/schemas/ReportFormat'
        options:
          type: object
          properties:
            includeCharts:
              type: boolean
              default: true
            includeMaps:
              type: boolean
              default: true
            includeAppendix:
              type: boolean
              default: false
            language:
              type: string
              default: en
            units:
              type: string
              enum: [metric, imperial]
              default: metric

    # Response schemas
    Pagination:
      type: object
      properties:
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
        hasMore:
          type: boolean
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
...
    Report:
      type: object
      properties:
        id:
          type: string
          format: uuid
        analysisId:
          type: string
          format: uuid
        name:
          type: string
        type:
          $ref: '#/components/schemas/ReportType'
        format:
          $ref: '#/components/schemas/ReportFormat'
        status:
          type: string
          enum: [GENERATING, COMPLETED, FAILED]
        fileSize:
          type: integer
        createdAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
    ReportListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/Report'
        pagination:
          $ref: '#/components/schemas/Pagination'
    ReportResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/Report'
        message:
          type: string
    ReportDetailResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          allOf:
            - $ref: '#/components/schemas/Report'
            - type: object
              properties:
                options:
                  type: object
                generatedAt:
                  type: string
                  format: date-time
    ReportDownloadResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            reportId:
              type: string
              format: uuid
            name:
              type: string
            format:
              $ref: '#/components/schemas/ReportFormat'
            fileSize:
              type: integer
            downloadUrl:
              type: string
              format: uri
            expiresAt:
              type: string
              format: date-time
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...